+++ lib/emque/consuming/configuration.rb
@@ -3,13 +3,14 @@ require "logger"
 module Emque
   module Consuming
     class Configuration
-      attr_accessor :app_name, :adapter, :error_handlers, :error_limit,
-        :error_expiration, :status, :status_port, :status_host, :socket_path,
-        :shutdown_handlers
+      attr_accessor :app_name, :auto_shutdown, :adapter, :error_handlers,
+        :error_limit, :error_expiration, :status, :status_port, :status_host,
+        :socket_path, :shutdown_handlers
       attr_writer :env, :log_level
 
       def initialize
-        @app_name = ""
+        @app_name           = ""
+        @auto_shutdown      = false
         @error_handlers     = []
         @error_limit        = 5
         @error_expiration   = 3600 # 60 minutes
@@ -41,6 +42,7 @@ module Emque
         {}.tap { |config|
           [
             :app_name,
+            :auto_shutdown,
             :adapter,
             :env,
             :error_handlers,